requirements.txt
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
tkinter-tooltip==2.1.0
matplotlib==3.7.2
python-telegram-bot==20.5
ta==0.10.2
websocket-client==1.6.3
schedule==1.2.0 